requirements.txt
pandas==2.1.3
langchain==0.0.340
sentence-transformers==2.2.2
shortuuid==1.0.11
PyYAML==6.0.1
